Add update mail method, rename
